+++ Emuera/Runtime/Script/Statements/Instraction.Child.cs
@@ -2824,16 +2824,16 @@ internal sealed partial class FunctionIdentifier
 			else
 				datFilename = soundArg.Str.GetStrValue(exm);
 			int repeat = soundArg.Opt != null ? (int)Math.Max(soundArg.Opt.GetIntValue(exm), 1) : 1;
-			string filepath = Path.GetFullPath(".\\sound\\" + datFilename);
+			string filepath = Program.MusicDir + datFilename;
 			try
 			{
-				if (File.Exists(filepath))
+				if (Program.FileExists(ref filepath))
 				{
 					int i;
 					for (i = 0; i < GlobalStatic.Sound.Length; i++)
 					{
 						if (GlobalStatic.Sound[i] == null)
-							GlobalStatic.Sound[i] = new Sound();
+							GlobalStatic.Sound[i] = Sound.Factory();
 						//未使用もしくは再生完了してる要素を使う
 						if (!GlobalStatic.Sound[i].isPlaying())
 							break;
@@ -2864,7 +2864,7 @@ internal sealed partial class FunctionIdentifier
 			for (int i = 0; i < GlobalStatic.Sound.Length; i++)
 			{
 				if (GlobalStatic.Sound[i] == null)
-					GlobalStatic.Sound[i] = new Sound();
+					GlobalStatic.Sound[i] = Sound.Factory();
 				if (GlobalStatic.Sound[i].isPlaying())
 					GlobalStatic.Sound[i].stop();
 			}
@@ -2928,7 +2928,7 @@ internal sealed partial class FunctionIdentifier
 			for (int i = 0; i < GlobalStatic.Sound.Length; i++)
 			{
 				if (GlobalStatic.Sound[i] == null)
-					GlobalStatic.Sound[i] = new Sound();
+					GlobalStatic.Sound[i] = Sound.Factory();
 				GlobalStatic.Sound[i].setVolume(vol);
 			}
 		}
@@ -2998,13 +2998,9 @@ internal sealed partial class FunctionIdentifier
 					}
 					if (version != GlobalStatic.GameBaseData.VersionName)
 					{
-						DialogResult result = MessageBox.Show(string.Format(trmb.NewVersionAvailable.Text, version, link),
-							trmb.UpdateCheck.Text,
-							MessageBoxButtons.YesNo,
-							MessageBoxIcon.None,
-							MessageBoxDefaultButton.Button2
-							);
-						if (result == DialogResult.Yes)
+						bool result = PlatformInterop.ShowQuestion(string.Format(trmb.NewVersionAvailable.Text, version, link),
+							trmb.UpdateCheck.Text);
+						if (result)
 						{
 							exm.VEvaluator.RESULT = 2;
 							System.Diagnostics.Process.Start(new System.Diagnostics.ProcessStartInfo