+++ Emuera/Runtime/Utils/Sys.cs
@@ -1,4 +1,4 @@
-﻿using System.Windows.Forms;
+using System.Windows.Forms;
 using System.IO;
 using System;
 
@@ -10,13 +10,13 @@ public static class AssemblyData
 	{
 		ExePath = Environment.ProcessPath;
 		#region eee_カレントディレクトリー
-		WorkingDir = Directory.GetCurrentDirectory() + "\\";
+		WorkingDir = Directory.GetCurrentDirectory() + Path.DirectorySeparatorChar;
 		#endregion
-		ExeDir = Path.GetDirectoryName(ExePath) + "\\";
+		ExeDir = Path.GetDirectoryName(ExePath) + Path.DirectorySeparatorChar;
 		ExeName = Path.GetFileName(ExePath);
 		emueraVer = typeof(AssemblyData).Assembly.GetName().Version;
 
-		EmueraVersionText = "Emuera.NET SkiaSharp " + Application.ProductVersion;
+		EmueraVersionText = "Emuera.NET SkiaSharp " + PlatformInterop.GetProductVersion();
 	}
 
 	/// <summary>