+++ Emuera/UI/Game/EmueraConsole.cs
@@ -671,7 +671,7 @@ internal sealed partial class EmueraConsole : IDisposable
 		RefreshStrings(true);
 		state = ConsoleState.Sleep;
 		process.UpdateCheckInfiniteLoopState();
-		Application.DoEvents();
+		PlatformInterop.DoEvents();
 		if (time > 0)
 			System.Threading.Thread.Sleep(time);
 		////DoEvents()の間にウインドウが閉じられたらおしまい。
@@ -906,7 +906,7 @@ internal sealed partial class EmueraConsole : IDisposable
 		//	wait_timeout = true;
 		//	while (countTime < timeLimit)
 		//	{
-		//		Application.DoEvents();
+		//		PlatformInterop.DoEvents();
 		//	}
 		//	wait_timeout = false;
 		//}
@@ -1360,7 +1360,7 @@ internal sealed partial class EmueraConsole : IDisposable
 				if (state != ConsoleState.WaitInput)
 					break;
 				//マクロループ時は待ち処理が起こらないのでここでシステムキューを捌く
-				Application.DoEvents();
+				PlatformInterop.DoEvents();
 #if DEBUG
 				if (state != ConsoleState.WaitInput || inputReq == null)
 					throw new ExeEE("");
@@ -1684,7 +1684,7 @@ internal sealed partial class EmueraConsole : IDisposable
 		{
 			while (_drawStopwatch.ElapsedMilliseconds < msPerFrame)
 			{
-				Application.DoEvents();
+				PlatformInterop.DoEvents();
 			}
 		}
 		window.TextBox.BackColor = bgColor.ToDrawingColor();
@@ -1981,6 +1981,7 @@ internal sealed partial class EmueraConsole : IDisposable
 			if (g.IsCreated)
 			{
 				SKBitmap img = g.SKBitmap;
+#if WINDOWS
 				try
 				{
 					e.Graphics.DrawImage(img.ToBitmap(), 0, 0);
@@ -1990,6 +1991,7 @@ internal sealed partial class EmueraConsole : IDisposable
 				{
 					// Fall through to default tooltip handling if conversion fails
 				}
+#endif
 			}
 
 		}